+++ TestUtilities/Mocks/NetworkInstrumentationMocks.swift
@@ -7,9 +7,13 @@
 import Foundation
 import DatadogInternal
 
-extension SpanID {
+extension SpanID: AnyMockable, RandomMockable {
     public static func mockAny() -> SpanID {
         return SpanID(rawValue: .mockAny())
+    }
+
+    public static func mockRandom() -> SpanID {
+        return SpanID(rawValue: .mockRandom())
     }
 
     public static func mock(_ rawValue: UInt64) -> SpanID {
@@ -18,9 +22,13 @@ extension SpanID {
 }
 
 
-extension TraceID {
+extension TraceID: AnyMockable, RandomMockable {
     public static func mockAny() -> TraceID {
         return TraceID(rawValue: (.mockAny(), .mockAny()))
+    }
+
+    public static func mockRandom() -> TraceID {
+        return TraceID(idHi: .mockRandom(), idLo: .mockRandom())
     }
 
     public static func mock(_ rawValue: (UInt64, UInt64)) -> TraceID {
@@ -33,6 +41,38 @@ extension TraceID {
 
     public static func mock(_ idLo: UInt64) -> TraceID {
         return TraceID(idLo: idLo)
+    }
+}
+
+extension TraceContext: AnyMockable, RandomMockable {
+    public static func mockAny() -> TraceContext {
+        return .mockWith()
+    }
+
+    public static func mockRandom() -> TraceContext {
+        return .mockWith(
+            traceID: .mockRandom(),
+            spanID: .mockRandom(),
+            parentSpanID: .mockRandom(),
+            sampleRate: .mockRandom(min: 0, max: 100),
+            isKept: .random()
+        )
+    }
+
+    public static func mockWith(
+        traceID: TraceID = .mockAny(),
+        spanID: SpanID = .mockAny(),
+        parentSpanID: SpanID? = nil,
+        sampleRate: Float = .mockAny(),
+        isKept: Bool = .mockAny()
+    ) -> TraceContext {
+        return TraceContext(
+            traceID: traceID,
+            spanID: spanID,
+            parentSpanID: parentSpanID,
+            sampleRate: sampleRate,
+            isKept: isKept
+        )
     }
 }
 
@@ -88,7 +128,7 @@ public final class URLSessionHandlerMock: DatadogURLSessionHandler {
     public var firstPartyHosts: FirstPartyHosts
 
     public var modifiedRequest: URLRequest?
-    public var parentSpan: TraceContext?
+    public var injectedTraceContext: TraceContext?
     public var shouldInterceptRequest: ((URLRequest) -> Bool)?
 
     public var onRequestMutation: ((URLRequest, Set<TracingHeaderType>) -> Void)?
@@ -111,13 +151,9 @@ public final class URLSessionHandlerMock: DatadogURLSessionHandler {
         interceptions.values.first { $0.request.url == url }
     }
 
-    public func modify(request: URLRequest, headerTypes: Set<TracingHeaderType>) -> URLRequest {
+    public func modify(request: URLRequest, headerTypes: Set<TracingHeaderType>) -> (URLRequest, TraceContext?) {
         onRequestMutation?(request, headerTypes)
-        return modifiedRequest ?? request
-    }
-
-    public func traceContext() -> TraceContext? {
-        parentSpan
+        return (modifiedRequest ?? request, injectedTraceContext)
     }
 
     public func interceptionDidStart(interception: URLSessionTaskInterception) {