+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Unreleased
 
 - [IMPROVEMENT] Start sending data immediately after SDK is initialized. See [#1798][]
+- [FEATURE] `DatadogTrace` now supports head-based sampling. See [#1794][]
+- [FEATURE] Support WebView recording in Session Replay. See [#1776][]
+- [IMPROVEMENT] Add `isInitialized` and `stopInstance` methods to ObjC API. See [#1800][]
+- [IMPROVEMENT] Add `addUserExtraInfo` method to ObjC API. See [#1799][]
+- [FIX] Use trace and span id as decimal. See [#1807][]
 
 # 2.10.0 / 23-04-2024
 
@@ -633,17 +638,22 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1697]: https://github.com/DataDog/dd-sdk-ios/pull/1697
 [#1707]: https://github.com/DataDog/dd-sdk-ios/pull/1707
 [#1711]: https://github.com/DataDog/dd-sdk-ios/pull/1711
+[#1721]: https://github.com/DataDog/dd-sdk-ios/pull/1721
 [#1722]: https://github.com/DataDog/dd-sdk-ios/pull/1722
 [#1724]: https://github.com/DataDog/dd-sdk-ios/pull/1724
 [#1741]: https://github.com/DataDog/dd-sdk-ios/pull/1741
 [#1742]: https://github.com/DataDog/dd-sdk-ios/pull/1742
 [#1746]: https://github.com/DataDog/dd-sdk-ios/pull/1746
+[#1747]: https://github.com/DataDog/dd-sdk-ios/pull/1747
+[#1794]: https://github.com/DataDog/dd-sdk-ios/pull/1794
 [#1774]: https://github.com/DataDog/dd-sdk-ios/pull/1774
 [#1763]: https://github.com/DataDog/dd-sdk-ios/pull/1763
 [#1767]: https://github.com/DataDog/dd-sdk-ios/pull/1767
 [#1798]: https://github.com/DataDog/dd-sdk-ios/pull/1798
 [#1721]: https://github.com/DataDog/dd-sdk-ios/pull/1721
 [#1747]: https://github.com/DataDog/dd-sdk-ios/pull/1747
+[#1776]: https://github.com/DataDog/dd-sdk-ios/pull/1776
+[#1807]: https://github.com/DataDog/dd-sdk-ios/pull/1807
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu